--- src/app/_components/layout/Header/HeaderActions/HeaderActions.module.scss
@@ -1,5 +1,0 @@
-.container {
-  margin-left: auto;
-  display: flex;
-  gap: 1rem;
-}
