+++ src/app/_components/layout/header/header-actions/account-menu/AccountMenu.module.scss
@@ -1,0 +1,167 @@
+.avatarButton {
+  padding: 0;
+  border-radius: 100%;
+  height: 45px;
+  width: 45px;
+  box-shadow: 0 2px 10px var(--black-a7);
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  &:focus {
+    box-shadow: 0 0 0 2px black;
+  }
+}
+
+.dropdownMenuContent,
+.dropdownMenuSubContent {
+  min-width: 250px;
+  background-color: white;
+  border-radius: 6px;
+  padding: 12px;
+  box-shadow:
+    0px 10px 38px -10px rgba(22, 23, 24, 0.35),
+    0px 10px 20px -15px rgba(22, 23, 24, 0.2);
+  animation-duration: 400ms;
+  animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
+  will-change: transform, opacity;
+}
+.dropdownMenuContent[data-side='top'],
+.dropdownMenuSubContent[data-side='top'] {
+  animation-name: slideDownAndFade;
+}
+.dropdownMenuContent[data-side='right'],
+.dropdownMenuSubContent[data-side='right'] {
+  animation-name: slideLeftAndFade;
+}
+.dropdownMenuContent[data-side='bottom'],
+.dropdownMenuSubContent[data-side='bottom'] {
+  animation-name: slideUpAndFade;
+}
+.dropdownMenuContent[data-side='left'],
+.dropdownMenuSubContent[data-side='left'] {
+  animation-name: slideRightAndFade;
+}
+
+.dropdownMenuItem,
+.dropdownMenuCheckboxItem,
+.dropdownMenuRadioItem,
+.dropdownMenuSubTrigger {
+  font-size: 18px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  line-height: 1;
+  border-radius: 3px;
+  height: 25px;
+  padding: 10px 5px;
+  position: relative;
+  padding-left: 25px;
+  user-select: none;
+  outline: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+.dropdownMenuSubTrigger[data-state='open'] {
+  background-color: var(--violet-4);
+  color: var(--violet-11);
+}
+.dropdownMenuItem[data-disabled],
+.dropdownMenuCheckboxItem[data-disabled],
+.dropdownMenuRadioItem[data-disabled],
+.dropdownMenuSubTrigger[data-disabled] {
+  color: var(--mauve-8);
+  pointer-events: none;
+}
+.dropdownMenuItem[data-highlighted],
+.dropdownMenuCheckboxItem[data-highlighted],
+.dropdownMenuRadioItem[data-highlighted],
+.dropdownMenuSubTrigger[data-highlighted] {
+  background-color: var(--green-11);
+  color: var(--violet-1);
+}
+
+.dropdownMenuLabel {
+  padding-left: 25px;
+  font-size: 12px;
+  line-height: 25px;
+  color: var(--mauve-11);
+}
+
+.dropdownMenuSeparator {
+  height: 1px;
+  background-color: var(--green-11);
+  margin: 10px;
+}
+
+.dropdownMenuItemIndicator {
+  position: absolute;
+  left: 0;
+  width: 25px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.dropdownMenuArrow {
+  fill: white;
+}
+
+.rightSlot {
+  margin-left: auto;
+  padding-left: 20px;
+  color: var(--mauve-11);
+}
+[data-highlighted] > .RightSlot {
+  color: white;
+}
+[data-disabled] .RightSlot {
+  color: var(--mauve-8);
+}
+
+@keyframes slideUpAndFade {
+  from {
+    opacity: 0;
+    transform: translateY(2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes slideRightAndFade {
+  from {
+    opacity: 0;
+    transform: translateX(-2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@keyframes slideDownAndFade {
+  from {
+    opacity: 0;
+    transform: translateY(-2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes slideLeftAndFade {
+  from {
+    opacity: 0;
+    transform: translateX(2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
